code reorganization: new _private module; updated rectangles,points functions declaration
