+++ cv3/color_spaces.py
@@ -27,7 +27,7 @@ Functions:
 from functools import partial
 import cv2
 import numpy as np
-from ._utils import type_decorator
+from ._private._color_spaces import _cvt_color
 
 __all__ = [
     'cvt_color', 'cvtColor',
@@ -40,14 +40,6 @@ __all__ = [
     'hsv2bgr', 'hsv2rgb',
     'rgb', 'bgr', 'rgba', 'bgra'
 ]
-
-
-@type_decorator
-def _cvt_color(img, code):
-    if code in (cv2.COLOR_GRAY2RGB, cv2.COLOR_GRAY2RGBA):
-        if img.ndim == 3 and img.shape[-1] != 1:
-            raise ValueError('Image must be grayscale (2 dims)')
-    return cv2.cvtColor(img, code=code)
 
 
 def cvt_color(img, code):
